+++ README.md
@@ -1,11 +1,11 @@
-# smuggle-env
+# trapdoor
 
-smuggle-env monkeypatches `ENV` so any secret values are available when asked for by name (i.e. `ENV['SOME_SECRET']`) but are hidden when requested in bulk (i.e. `ENV.to_h`, `ENV.inspect`, iterators). This was designed to be a drop in replacement for codebases that use `ENV` to hold secrets but also use developer tools that may cause these values to be accidentally logged. 
+trapdoor monkeypatches `ENV` so any secret values are available when asked for by name (i.e. `ENV['SOME_SECRET']`) but are hidden when requested in bulk (i.e. `ENV.to_h`, `ENV.inspect`, iterators). This was designed to be a drop in replacement for codebases that use `ENV` to hold secrets but also use developer tools that may cause these values to be accidentally logged. 
 
 ## Installation
 
-1. Add to your Gemfile and install with bundle, or `gem install smuggle-env`
-2. `require 'smuggle-env'`
+1. Add to your Gemfile and install with bundle, or `gem install trapdoor`
+2. `require 'trapdoor'`
 
 ## Walkthrough
 
@@ -16,7 +16,7 @@ Let's start by inspecting the environment in the repl. We can see `SECRET_API_TO
  => {... "SECRET_API_TOKEN"=>"987tfghjo0987yt"}
 ```
 
-After loading `smuggle-env`, let's tell it to hide this specific value then inspect the environment again.
+After loading `trapdoor`, let's tell it to hide this specific value then inspect the environment again.
 
 ```
 > ENV.hide "SECRET_API_TOKEN"
