+++ lib/trapdoor.rb
